feat: 🚀 support DF 50.xx
Updates dependencies to support the new Dwarf Fortress version 50.xx. This comes with some changes to the app itself, because there are not raws stored with saves anymore. So simply select the install folder for Dwarf Fortress and let the app read vanilla, installed_mods and any workshop mods.

App no longer works with Dwarf Fortress versions prior to 50.xx (because of the changes to storing raw files outside of the saves).